chore: restructure applepay config parser
